+++ test/CMakeLists.txt
@@ -20,4 +20,4 @@ endif()
 target_link_libraries(any_sender_of_test PUBLIC gmock)
 target_link_libraries(async_manual_reset_event_test PUBLIC gmock)
 target_link_libraries(async_scope_test PUBLIC gmock)
-target_link_libraries(async_scope_v1_test PUBLIC gmock)
+target_link_libraries(async_scope_v0_test PUBLIC gmock)
